Break up WinAPI.cs into 3 different files. Create new Eutherion.Win.Native namespace.
