+++ Eutherion/Win/Eutherion.Win.csproj
@@ -65,6 +65,9 @@
     <Compile Include="ITextProvider.cs" />
     <Compile Include="LineSegment.cs" />
     <Compile Include="LiveTextFile.cs" />
+    <Compile Include="Native\NativeMethods.cs" />
+    <Compile Include="Native\ResizeMode.cs" />
+    <Compile Include="Native\Structures.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Storage\CompactSettingWriter.cs" />
     <Compile Include="Storage\FileNameType.cs" />
@@ -100,7 +103,6 @@
     <Compile Include="UIActions\UIMenu.cs" />
     <Compile Include="Utils\KeyUtilities.cs" />
     <Compile Include="Utils\ThemeHelper.cs" />
-    <Compile Include="WinAPI.cs" />
     <Compile Include="WorkingCopyTextFile.cs" />
   </ItemGroup>
   <ItemGroup>
