+++ Eutherion/Win/Native/ResizeMode.cs
@@ -1,6 +1,6 @@
 ﻿#region License
 /*********************************************************************************
- * WinAPI.cs
+ * ResizeMode.cs
  *
  * Copyright (c) 2004-2020 Henk Nicolai
  *
@@ -19,26 +19,8 @@
 **********************************************************************************/
 #endregion
 
-using System;
-using System.Diagnostics;
-using System.Runtime.InteropServices;
-using System.Security;
-
-namespace Eutherion.Win
+namespace Eutherion.Win.Native
 {
-    /// <summary>
-    /// Encapsulates the RECT structure which is used by the Windows API.
-    /// </summary>
-    [StructLayout(LayoutKind.Sequential)]
-    [DebuggerDisplay("({Left},{Top})-({Right},{Bottom})")]
-    public struct RECT
-    {
-        public int Left;
-        public int Top;
-        public int Right;
-        public int Bottom;
-    }
-
     /// <summary>
     /// Specifies the mode in which a form is being resized.
     /// </summary>
@@ -107,64 +89,5 @@ namespace Eutherion.Win
         /// This enumeration value is equal to WMSZ_BOTTOM + WMSZ_RIGHT.
         /// </remarks>
         BottomRight = 8,
-    }
-
-    /// <summary>
-    /// Encapsulates the ICONINFO structure which is used by the Windows API.
-    /// </summary>
-    [StructLayout(LayoutKind.Sequential)]
-    internal struct ICONINFO
-    {
-        public bool fIcon;
-        public int xHotspot;
-        public int yHotspot;
-        public IntPtr hbmMask;
-        public IntPtr hbmColor;
-    }
-
-    [StructLayout(LayoutKind.Sequential)]
-    public struct COPYDATASTRUCT
-    {
-        public IntPtr dwData;
-        public int cbData;
-        public IntPtr lpData;
-    }
-
-    /// <summary>
-    /// Contains P/Invoke definitions for the Windows API.
-    /// </summary>
-    [SuppressUnmanagedCodeSecurity]
-    public static class WinAPI
-    {
-        const string Gdi32 = "gdi32.dll";
-
-        [DllImport(Gdi32, CharSet = CharSet.Auto, ExactSpelling = true, SetLastError = true)]
-        public static extern bool DeleteObject(HandleRef hObject);
-
-        const string User32 = "user32.dll";
-
-        [DllImport(User32, CharSet = CharSet.Auto, ExactSpelling = true)]
-        internal static extern IntPtr CreateIconIndirect([In] ref ICONINFO iconInfo);
-
-        [DllImport(User32, CharSet = CharSet.Auto, ExactSpelling = true, SetLastError = true)]
-        public static extern bool DestroyIcon(HandleRef hIcon);
-
-        [DllImport(User32, CharSet = CharSet.Auto, ExactSpelling = true)]
-        public static extern IntPtr GetFocus();
-
-        [DllImport(User32, CharSet = CharSet.Auto, ExactSpelling = true)]
-        internal static extern bool GetIconInfo(HandleRef hIcon, ref ICONINFO info);
-
-        [DllImport(User32, CharSet = CharSet.Auto, ExactSpelling = true)]
-        public static extern bool HideCaret(HandleRef hWnd);
-
-        [DllImport(User32, CharSet = CharSet.Auto, SetLastError = true)]
-        public static extern IntPtr SendMessage(HandleRef hWnd, int msg, IntPtr wParam, IntPtr lParam);
-
-        [DllImport(User32, CharSet = CharSet.Auto, ExactSpelling = true)]
-        public static extern bool ShowCaret(HandleRef hWnd);
-
-        [DllImport(User32, CharSet = CharSet.Auto, ExactSpelling = true)]
-        public static extern bool ShowWindow(HandleRef hWnd, int nCmdShow);
     }
 }
