+++ Eutherion/Win/Native/Structures.cs
@@ -1,0 +1,64 @@
+﻿#region License
+/*********************************************************************************
+ * Structures.cs
+ *
+ * Copyright (c) 2004-2020 Henk Nicolai
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ *
+**********************************************************************************/
+#endregion
+
+using System;
+using System.Diagnostics;
+using System.Runtime.InteropServices;
+
+namespace Eutherion.Win.Native
+{
+    /// <summary>
+    /// Encapsulates the RECT structure which is used by the Windows API.
+    /// </summary>
+    [StructLayout(LayoutKind.Sequential)]
+    [DebuggerDisplay("({Left},{Top})-({Right},{Bottom})")]
+    public struct RECT
+    {
+        public int Left;
+        public int Top;
+        public int Right;
+        public int Bottom;
+    }
+
+    /// <summary>
+    /// Encapsulates the ICONINFO structure which is used by the Windows API.
+    /// </summary>
+    [StructLayout(LayoutKind.Sequential)]
+    public struct ICONINFO
+    {
+        public bool fIcon;
+        public int xHotspot;
+        public int yHotspot;
+        public IntPtr hbmMask;
+        public IntPtr hbmColor;
+    }
+
+    /// <summary>
+    /// Encapsulates the COPYDATASTRUCT structure which is used by the Windows API.
+    /// </summary>
+    [StructLayout(LayoutKind.Sequential)]
+    public struct COPYDATASTRUCT
+    {
+        public IntPtr dwData;
+        public int cbData;
+        public IntPtr lpData;
+    }
+}
