+++ Sandra.UI/AppTemplate/SingleInstanceMainForm.cs
@@ -19,6 +19,7 @@
 **********************************************************************************/
 #endregion
 
+using Eutherion.Win.Native;
 using System;
 using System.ComponentModel;
 using System.Runtime.InteropServices;
@@ -148,7 +149,7 @@ namespace Eutherion.Win.AppTemplate
                     Marshal.StructureToPtr(copyData, copyDataBuffer, false);
 
                     // SendMessage only returns when the receiving process has completed the call.
-                    WinAPI.SendMessage(remoteWindowHandle, WM_COPYDATA, IntPtr.Zero, copyDataBuffer);
+                    NativeMethods.SendMessage(remoteWindowHandle, WM_COPYDATA, IntPtr.Zero, copyDataBuffer);
 
                     int error = Marshal.GetLastWin32Error();
                     if (error != 0) throw new Win32Exception(error);
